firsdt running refactoring to new system

+++ include/ESP32_treadmill_tacho_config.h
@@ -106,7 +106,7 @@ struct TreadmillMetrics {
     float mps = 0;
     float mpsOffset = 0;
     uint8_t cpuUsagePercent = 0;   // CPU usage percentage (0-100)
-    uint32_t workoutDistance = 0;  // in mm
+    float workoutDistance = 0.0f;  // in meters (changed from mm)
     unsigned long workoutStartTime = 0;
     unsigned long lastUpdate = 0;
